backend/requirements.txt
annotated-types==0.7.0
anyio
bcrypt
black
boto3
botocore
certifi
cffi
charset-normalizer
click
cryptography
dnspython
ecdsa
email-validator
fastapi
flake8
h11
httpcore
httpx
idna
iniconfig
isort
jmespath
jq
markdown-it-py
mccabe
mdurl
motor
mypy
mypy_extensions
numpy
oauthlib
packaging
pandas
passlib
pathspec
platformdirs
pluggy
pyasn1
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
pytokens
pytz
requests
requests-oauthlib
rich
rsa
s3transfer
s5cmd
shellingham
six
sniffio
typer
tzdata
urllib3
uvicorn
watchfiles
reportlab
openpyxl
arabic-reshaper
python-bidi
